requirements.txt
faiss-cpu
transformers
Pillow
pypdf
jq
anthropic
langchain
langchain_community
langchain_core
streamlit
botocore
boto3
awscli
requests
requests-aws4auth
opensearch-py
python-dotenv==1.0.0
gradio>=4.0.0
pypdfium2
nltk
PyMuPDF
asyncio
pre-commit
ruff
